+++ .env.example
@@ -1,19 +1,33 @@
-MONGO_URI=mongodb://127.0.0.1:27017
-MONGO_DB=meowerserver
-REDIS_URI=redis://127.0.0.1:6379/0
-REAL_IP_HEADER=
-IPHUB_KEY=
-CL3_HOST="0.0.0.0"
-CL3_PORT=3000
+# REST API
+
 API_HOST="0.0.0.0"
 API_PORT=3001
 API_ROOT=
 
-CAPTCHA_SITEKEY=
-CAPTCHA_SECRET=
+API_INTERNAL_TOKEN=
+API_INTERNAL_URL=
+
+# Cloudlink
+
+CL3_HOST="0.0.0.0"
+CL3_PORT=3000
+
+# Database
+
+MONGO_DB=meowerserver
+MONGO_URI=mongodb://127.0.0.1:27017
+REDIS_URI=redis://127.0.0.1:6379/0
+
+# gRPC Services
 
 GRPC_AUTH_ADDRESS="0.0.0.0:5000"
 GRPC_AUTH_TOKEN=
-
 GRPC_UPLOADS_ADDRESS=
 GRPC_UPLOADS_TOKEN=
+
+# External Services
+
+CAPTCHA_SECRET=
+CAPTCHA_SITEKEY=
+IPHUB_KEY=
+REAL_IP_HEADER=
