+++ cloudlink/types.py
@@ -1,0 +1,21 @@
+api_error_statuscode_map = {
+	"accountBanned": "Banned",
+	"accountDeleted": "Deleted",
+	"badRequest": "Syntax",
+	"ipBlocked": "Blocked",
+	"mfaRequired": "2FARequired",
+	"registrationBlocked": "Blocked",
+	"tooManyRequests": "RateLimit",
+	"Unauthorized": "PasswordInvalid",
+}
+
+statuscodes = {
+	"Banned": "E:018 | Account Banned",
+	"Datatype": "E:102 | Datatype",
+	"InternalServerError": "E:104 | Internal",
+	"Invalid": "E:118 | Invalid command",
+	"OK": "I:100 | OK",
+	"PasswordInvalid": "I:011 | Invalid Password",
+	"Syntax": "E:101 | Syntax",
+	"TAEnabled": "I:112 | Trusted Access enabled",
+}
